Add missing txindex
W/o txindex on a fresh synced server, blockbook gives a transaction not found error
for old transaction.
This seems to be related to different code path in blockbook for bulk load and single block load,
since more recent transactions are still found.
With txindex the same old transactions are found.
(Default setting for bitcoin and others too).